fixed results page routing and changed detailed quiz result prompt
