+++ src/App.css
@@ -268,7 +268,7 @@ body {
   margin: 10px;
   padding: 10px;
   width: 400px;
-  height: 500px;
+  height: auto;
   background-color: #FFEECC;
   font-family: "Palatino";
   font-size: 15px;
